# chenzekai_kama/KamaChat/api/v1/session_controller.go
package v1

import (
	"github.com/gin-gonic/gin"
	"kama_chat_server/internal/dto/request"
	"kama_chat_server/internal/service/gorm"
	"kama_chat_server/pkg/constants"
	"kama_chat_server/pkg/zlog"
	"net/http"
)

// OpenSession 打开会话
// @Summary 打开会话
// @Tags session
// @Accept json
// @Produce json
// @Param body body request.OpenSessionRequest true "打开会话请求"
// @Success 200 {object} SwaggerResponse
// @Router /session/openSession [post]
func OpenSession(c *gin.Context) {
	var openSessionReq request.OpenSessionRequest
	if err := c.BindJSON(&openSessionReq); err != nil {
		zlog.Error(err.Error())
		c.JSON(http.StatusOK, gin.H{
			"code":    500,
			"message": constants.SYSTEM_ERROR,
		})
		return
	}
	message, sessionId, ret := gorm.SessionService.OpenSession(openSessionReq)
	JsonBack(c, message, ret, sessionId)
}

// GetUserSessionList 获取用户会话列表
// @Summary 获取用户会话列表
// @Tags session
// @Accept json
// @Produce json
// @Param body body request.OwnlistRequest true "用户ID"
// @Success 200 {object} SwaggerResponse
// @Router /session/getUserSessionList [post]
func GetUserSessionList(c *gin.Context) {
	var getUserSessionListReq request.OwnlistRequest
	if err := c.BindJSON(&getUserSessionListReq); err != nil {
		zlog.Error(err.Error())
		c.JSON(http.StatusOK, gin.H{
			"code":    500,
			"message": constants.SYSTEM_ERROR,
		})
		return
	}
	message, sessionList, ret := gorm.SessionService.GetUserSessionList(getUserSessionListReq.OwnerId)
	JsonBack(c, message, ret, sessionList)
}

// GetGroupSessionList 获取群聊会话列表
// @Summary 获取群聊会话列表
// @Tags session
// @Accept json
// @Produce json
// @Param body body request.OwnlistRequest true "用户ID"
// @Success 200 {object} SwaggerResponse
// @Router /session/getGroupSessionList [post]
func GetGroupSessionList(c *gin.Context) {
	var getGroupListReq request.OwnlistRequest
	if err := c.BindJSON(&getGroupListReq); err != nil {
		zlog.Error(err.Error())
		c.JSON(http.StatusOK, gin.H{
			"code":    500,
			"message": constants.SYSTEM_ERROR,
		})
		return
	}
	message, groupList, ret := gorm.SessionService.GetGroupSessionList(getGroupListReq.OwnerId)
	JsonBack(c, message, ret, groupList)
}

// DeleteSession 删除会话
// @Summary 删除会话
// @Tags session
// @Accept json
// @Produce json
// @Param body body request.DeleteSessionRequest true "删除会话请求"
// @Success 200 {object} SwaggerResponse
// @Router /session/deleteSession [post]
func DeleteSession(c *gin.Context) {
	var deleteSessionReq request.DeleteSessionRequest
	if err := c.BindJSON(&deleteSessionReq); err != nil {
		zlog.Error(err.Error())
		c.JSON(http.StatusOK, gin.H{
			"code":    500,
			"message": constants.SYSTEM_ERROR,
		})
		return
	}
	message, ret := gorm.SessionService.DeleteSession(deleteSessionReq.OwnerId, deleteSessionReq.SessionId)
	JsonBack(c, message, ret, nil)
}

// CheckOpenSessionAllowed 检查是否可以打开会话
// @Summary 检查是否可以打开会话
// @Tags session
// @Accept json
// @Produce json
// @Param body body request.CreateSessionRequest true "会话检查请求"
// @Success 200 {object} SwaggerResponse
// @Router /session/checkOpenSessionAllowed [post]
func CheckOpenSessionAllowed(c *gin.Context) {
	var req request.CreateSessionRequest
	if err := c.BindJSON(&req); err != nil {
		zlog.Error(err.Error())
		c.JSON(http.StatusOK, gin.H{
			"code":    500,
			"message": constants.SYSTEM_ERROR,
		})
		return
	}
	message, res, ret := gorm.SessionService.CheckOpenSessionAllowed(req.SendId, req.ReceiveId)
	JsonBack(c, message, ret, res)
}