# chenzekai_kama/KamaChat/docs/swagger.yaml
basePath: /
definitions:
...
  request.CreateSessionRequest:
    properties:
      receive_id:
        type: string
      send_id:
        type: string
    type: object
...
  request.DeleteSessionRequest:
    properties:
      owner_id:
        type: string
      session_id:
        type: string
    type: object
...
  request.OpenSessionRequest:
    properties:
      receive_id:
        type: string
      send_id:
        type: string
    type: object
  request.OwnlistRequest:
    properties:
      owner_id:
        type: string
    type: object
...
  v1.SwaggerResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: ok
        type: string
    type: object
info:
  contact: {}
  description: KamaChat server API documentation.
  title: KamaChat API
  version: "1.0"
paths:
...
  /session/checkOpenSessionAllowed:
    post:
      consumes:
      - application/json
      parameters:
      - description: 会话检查请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CreateSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 检查是否可以打开会话
      tags:
      - session
  /session/deleteSession:
    post:
      consumes:
      - application/json
      parameters:
      - description: 删除会话请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.DeleteSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 删除会话
      tags:
      - session
  /session/getGroupSessionList:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.OwnlistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 获取群聊会话列表
      tags:
      - session
  /session/getUserSessionList:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.OwnlistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 获取用户会话列表
      tags:
      - session
  /session/openSession:
    post:
      consumes:
      - application/json
      parameters:
      - description: 打开会话请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.OpenSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 打开会话
      tags:
      - session
...
schemes:
- https
- http
swagger: "2.0"